Major work on overhauling the system to work in a more distributed manner.
Refactored all of the RPC callback code for the HStoreCoordinatorNode

The TransactionState now keeps track of FragmentResponses as well as VoltTable results coming in from the remote partitions. This ensures that there isn't a race condition where we try to run some new fragments in the Dtxn.Coordinator but before the responses come back.

Finally cleaned up the FragmentTask and FragmentResponse messages stuff. We can now attach dependency ids but without having to actually store the data (which is passed through HStoreMessenger separately).

This is obviously all still a work in progress so tread lightly kids.


git-svn-id: https://database.cs.brown.edu/svn/hstore/branches/protorpc-messaging-branch@1621 7a8f3d5b-8c5f-44cb-9125-2414a256df87