HUGE WIN! STRAP ON YOUR HATS BECAUSE THIS IS A GOOD ONE!
H-Store now supports SPECULATIVE EXECUTION!!!
We can use the Markov models to figure when we are done with partitions and then enable speculative execution.
If know that a transaction is read-only at a partition when in spec exec mode, we can let any read-only txns commit right away.
Going home to have dinner and then test on the Wisconsin cluster.

I'd like to dedicate this commit to little KB. She made me soup this weekend.
I'd also like to dedicate it to Evan. I've seem him help a hobo and knock down an old lady in a single day!


git-svn-id: https://database.cs.brown.edu/svn/hstore/branches/markov-branch@2054 7a8f3d5b-8c5f-44cb-9125-2414a256df87