Major overhaul of how we manage transactions at each ExecutionSite: + Moved all of the internal data structures for each partition from HStoreSite into ExecutionSite + Each ExecutionSite is now responsible for determining whether transactions should be speculatively executed and when they can return their results to the clients. + We can now speculatively execute single-partition transactions as they come in, and not just after we're told that the multi-partition is done at a partition. + Removed superfluous synchronization locks in ExecutionSite
I also fixed up how we maintain the Dtxn.Coordinator's configuration file. Previously we used to have to create a separate file from the catalog jar. Now the VoltCompiler embeds the file into the jar, and the 'dtxn-coordinator' ant target unpacks it at runtime. Removes the need to call 'hstore-conf'

Started working on a cleaner HStoreConf too. Eventually we'll merge this with the properties file.

KB is making me dinner right now, so I'm going to clean things up a bit before then...


git-svn-id: https://database.cs.brown.edu/svn/hstore/branches/markov-branch@2137 7a8f3d5b-8c5f-44cb-9125-2414a256df87