More refactoring. Not sure why I did it, but oh well it's better now. + QueryPlanNodeUtil has been completely moved into PlanNodeUtil. This will cut down on the number of duplicate methods + Renamed PlanNodeUtil.getOutputColumns() to PlanNodeUtil.getUpdatedColumns(). Then I wrote a legit PlanNodeUtil.getOutputColumns() + DesignerHints uses reflection to get the fields to serialize, instead of having to maintain a large Enum + Moved stuff from CatalogUtil to the appropriate utility class + Moved out all of the static methods from AbstractPartitioner into PartitionerUtil
As for new work, I started adding support for vertical partitioning. First step is to get the candidate sets of columns that we could put into a vertical partition. Still a work in progress.
I also started working on a new deployment script to launch a cluster in EC2 using Boto. Still a work in progress.


git-svn-id: https://database.cs.brown.edu/svn/hstore/branches/partitioning-branch@2240 7a8f3d5b-8c5f-44cb-9125-2414a256df87