Finished coding AuctionMark 2.0 + Consolidated CheckWinningBids and PostAuction into a single transaction called ClosedAuction. It does the same operations as those two old ones + Cleaned up the parameter generating logic in AuctionMarkClient + Tried to optimize various SQL statements in some of the transactions.
There is still a bug where we keep calling NewPurchase on the same items. I need to fix this.
I also need to test on larger clusters and with multiple clients. Going to head to head home and do that now.
KB left for College Park today so I'm on my own...


git-svn-id: https://database.cs.brown.edu/svn/hstore/branches/partitioning-branch@2281 7a8f3d5b-8c5f-44cb-9125-2414a256df87