A couple of things here:
(1) We now only build the anticaching component in the EE if the HStoreConf 'site.anticache_enable' feature is enabled. This was for @jdebrabant

(2) Likewise, you can now pass in from the commandline whether you want the reversible LRU chain in the anti-cache

(3) Added initial code to get @jarulraj started on building out the mmap version of H-Store's storage manager.