Update DB Schema to include additional tables and stream for nearby stations and discount
Changed the data type of current_discount in StationStatus to Integer instead of
Float as it is a count rather than the actual value.
Added s1 stream to hold the cascade user_id from the bikeStatus stream.
Added nearByStations and nearByDiscounts tables.  The nearByStation is to hold the last N
near by stations for a given user.  The nearByDiscount is to hold the last N stations with
available discounts.