There are a couple of things going on in here related to #164
First, the EE no longer tries to set a dummy partitionId in the EvictedTupleAccessException. I actually think that this is unnecessary because there is another problem with serializing the exception.
I think that we are trying to do something in the EE in a non-thread safe manner. I'm not sure what, but I think we are blowing away the EE's shared result buffer, which causes the deserialization of the EvictedTupleAccessException to get messed up.
I also commented out AntiCacheEvictionManager::readBlocks(). This method does not appear to be called by anything. It is probably safe to remove entirely.

This also has the new check to make sure that we don't try to access an evicted tuple that is marked as deleted.